avgpool: use correct requant datatypes

+++ XpulpNN/64bit/include/pulp_nn_kernels.h
@@ -3474,8 +3474,8 @@ void pulp_nn_avgpool_u8_u8(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3496,8 +3496,8 @@ void pulp_nn_avgpool_u8_u4(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3518,8 +3518,8 @@ void pulp_nn_avgpool_u8_u2(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3540,8 +3540,8 @@ void pulp_nn_avgpool_u4_u8(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3562,8 +3562,8 @@ void pulp_nn_avgpool_u4_u4(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3584,8 +3584,8 @@ void pulp_nn_avgpool_u4_u2(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3606,8 +3606,8 @@ void pulp_nn_avgpool_u2_u8(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3628,8 +3628,8 @@ void pulp_nn_avgpool_u2_u4(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
@@ -3650,8 +3650,8 @@ void pulp_nn_avgpool_u2_u2(
     uint16_t dim_im_out_x,
     uint16_t dim_im_out_y,
     uint16_t out_shift,
-    uint32_t out_add,
-    uint32_t lambda,
+    int64_t out_add,
+    int64_t lambda,
     uint8_t * Im_out,
     int flag_requant
 );
